Improve Vulkan byte range descriptions
Renames the "cont.d" column in the Pipeline State window to "View".
This is more accurate for all resource types except samplers, in which
case the contents is self-explanatory.

Adds a "(whole size)" annotation after the range if VK_WHOLE_SIZE was
used, and an "(empty view)" annotation if the range is empty. The
latter annotation is helpful for diagnosing incorrectly filled buffer
descriptors.