Improve Vulkan resource descriptions
Removes the "cont.d" column in the Pipeline State window Resources list
and consolidates it into "Contents". The removed column had a confusing
name and displayed different data depending on the resource type.

Adds a "(whole size)" annotation after the range if VK_WHOLE_SIZE was
used, and an "(empty view)" annotation if the range is empty. These
annotations are helpful for diagnosing incorrectly filled buffer
descriptors.